requirements_mac_m1_m2_m4.txt
tensorflow-macos
tensorflow-metal
fastapi
python-multipart
uvicorn[standard]
pillow
numpy
scipy